refactor(cli): introduce hive data layer lib
