+++ packages/libraries/cli/src/helpers/hive/schema.ts
@@ -1,0 +1,1 @@
+export * as Schema from '../../gql/graphql';
